Using tcp passthrough mode for TLS
The flexibility of Traefik permit to switch from reverse proxy mode to
passthrough pretty easily as seen in docker-compose we only need to
change some labels.

However some other changes are needed because certificates are now
located on the server while before they were located on the ingress.
As a consequence we need splitted DNS rules for the CA (because of ACME)

Keep both docker compose files for completeness